Add omrthread_get_self_thread_time()
This function returns the user and system cpu time of the calling
thread. This new function is needed because the only way that we're
aware of to get the user CPU time of a thread is to use
getrusage(RUSAGE_THREAD, ...) which can only be called on the current
thread.

Related: https://github.com/eclipse-openj9/openj9/pull/20186

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>